linux-yocto/5.15: fix qemuarm graphical boot
During the cleanup/refactoring of qemuarm* we dropped a PCI option
that is required for graphical boot. The configuration is fixed to
create a separate fragment, which just enabled the minimim and we
include it into qemuarma15 standard/preempt-rt.

Integrating the following commit(s) to linux-yocto/5.15:

    fcf48627ea5 qemuarma15: include pci-of-generic support

(From OE-Core rev: 3387700629a479f3e6674830e0c53d8a9e67b2c6)

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>